requirements.txt
ultralytics>=8.0.0
streamlit
pillow
opencv-python
numpy
